Separate indexer from other utils
